refactor display modal show product

+++ shop/public/admin/assets/base/product.js
@@ -1,12 +1,12 @@
-// delete data when hidden any modal
+
 $(document).ready(function () {
     $('.modal').on('hidden.bs.modal', function () {
         var inputs = document.querySelectorAll('.modal-body input');
         inputs.forEach(function(input) {
             input.value = '';
         });
-    document.getElementById('show-image-create').src = '';
-    $('#productTableBody').empty();
+        document.getElementById('show-image-create').src = '';
+        $('#productTableBody').empty();
     });
 });
 
@@ -46,23 +46,24 @@ function displayShowModal(product_id) {
             success: function(response) {
                 if (response.hasOwnProperty('product')) {
                     var product = response.product;
+
+                    // Tạo HTML cho danh sách các categories
+                    var categoriesHtml = "";
+                    for (var i = 0; i < product.categories.length; i++) {
+                        categoriesHtml += `${product.categories[i].name}<br>`;
+                    }
                     
-                        // Tạo HTML cho danh sách các categories
-                        var categoriesHtml = "";
-                        for (var j = 0; j < product.categories.length; j++) {
-                            categoriesHtml += product.categories[j].name + "<br>";
-                        }
-                    
-                        // Tạo HTML cho từng sản phẩm
-                        var productHTML = '<tr>' +
-                                              '<td>' + product.id + '</td>' +
-                                              '<td>' + (product.images.length > 0 ? '<img src="upload/' + product.images[0].url + '" alt="Product Image" width="200px" height="200px">' : '<h4>#</h4>') + '</td>' +
-                                              '<td>' + product.name + '</td>' +
-                                              '<td>' + product.price + '</td>' +
-                                              '<td>' + product.sale + '</td>' +
-                                              '<td>' + categoriesHtml + '</td>' +
-                                              '<td>' + product.description + '</td>' +
-                                          '</tr>';
+                    // Tạo HTML cho từng sản phẩm
+                    var productHTML = `<tr>
+                                          <td>${product.id}</td>
+                                          <td>${product.images.length > 0 ? `<img src="upload/${product.images[0].url}" alt="Product Image" width="200px" height="200px">` : '<h4>#</h4>'}</td>
+                                          <td>${product.name}</td>
+                                          <td>${product.price}</td>
+                                          <td>${product.sale}</td>
+                                          <td>${categoriesHtml}</td>
+                                          <td>${product.description}</td>
+                                      </tr>`;
+
                         // Thêm sản phẩm vào modal show
                         $('#showModal #productTableBody').append(productHTML);
                     
@@ -71,7 +72,7 @@ function displayShowModal(product_id) {
                 }
             },
             error: function(error) {
-                console.log("Error loading product data: " + error);
+                alert("Error " + error);
             }
           });
 }
@@ -82,10 +83,8 @@ function displayEditModal(product_id){
         url: "/products/" + product_id+ "/edit",
         method: "GET",
         success: function(response) {
-            if (response.hasOwnProperty('error')) {
-                console.error("Error loading product data: " + response.error);
-            } else if (response.hasOwnProperty('product')) {
-
+            if (response.hasOwnProperty('product')) {
+                
                 var product = response.product;
                 $("#edit-id").val(product.id);
                 $("#edit-name").val(product.name);
@@ -112,8 +111,8 @@ function displayEditModal(product_id){
                 });
                 $("#editModal").modal("show");
             }
-        },error: function (result) {
-            alert('loi');
+        },error: function (response) {
+            alert('Error' + response.message);
         }
     });
 }
@@ -133,12 +132,12 @@ function submitEditModal() {
         data: formData,
         contentType: false, 
         processData: false,
-        success: function(result) {
-            alert(result.message);
+        success: function(response) {
+            alert('Success' + response.message);
             location.reload();
         },
-        error: function (result) {
-            var errors = result.responseJSON.errors;
+        error: function (response) {
+            var errors = response.responseJSON.errors;
         // Lặp qua từng trường và hiển thị thông báo lỗi
             for (var field in errors) {
                 if (errors.hasOwnProperty(field)) {
@@ -162,46 +161,53 @@ function searchProduct() {
             searchInput: searchInput
         },
         success: function(response) {
-            
-            var products = response.product.data;
-            var roleUsers = response.roles.map(role => role.name);
-           
-            $('#searchProductsIndex').empty(); // xoa du lieu cua view index
-            $('#paginationLinks').hide();
-            for (var i = 0; i < products.length; i++) {
-                var product = products[i];
-                var productHTML = `
-                    <tr>
-                        <td>${product.id}</td>
-                        <td>${product.images.length > 0 ? `<img src="upload/${product.images[0].url}" alt="Product Image" width="100px" height="100px">` : '<h4>#</h4>'}</td>
-                        <td>${product.name}</td>
-                        <td>${product.price}</td>
-                        <td>${product.sale}</td>
-                        <td>${product.categories.map(category => category.name).join('<br>')}</td>
-                        <td>
-                            <div style="display: flex;">
-                                <button type="button" class="btn btn-success" style="margin-left: 5px;" data-bs-toggle="modal" data-bs-target="#showModal" onclick="displayShowModal(${product.id})">
-                                    <i class="fas fa-eye"></i>
-                                </button>
-                                ${roleUsers.some(role => ['admin','super-admin'].includes(role)) ? `
-                                    <button type="button" class="btn btn-warning" style="margin-left: 5px;" data-bs-toggle="modal" data-bs-target="#editModal" onclick="displayEditModal(${product.id})">
-                                        <i class="fas fa-edit"></i>
-                                    </button>` : ''
-                                }
-                                ${roleUsers.some(role => ['super-admin'].includes(role)) ? `
-                                    <button type="button" class="btn btn-danger" style="margin-left: 5px;" onclick="handleDelete(${product.id})">
-                                        <i class="fas fa-trash-alt"></i>
-                                    </button>` : ''
-                                }
-                                
-                            </div>
-                        </td>
-                    </tr>
-                `;
-              
-                $('#searchProductsIndex').append(productHTML); // show du lieu len view index
+            if (response.hasOwnProperty('product')) {
+                var products = response.product.data;
+                var roleUsers = response.roles.map(role => role.name);
+                productsLength = products.length;
+                $('#searchProductsIndex').empty(); // xoa du lieu cua view index
+                $('#paginationLinks').hide();
+                if(productsLength >0){
+                    for (var i = 0; i < productsLength; i++) {
+                        var product = products[i];
+                        var productHTML = `
+                            <tr>
+                                <td>${product.id}</td>
+                                <td>${product.images.length > 0 ? `<img src="upload/${product.images[0].url}" alt="Product Image" width="100px" height="100px">` : '<h4>#</h4>'}</td>
+                                <td>${product.name}</td>
+                                <td>${product.price}</td>
+                                <td>${product.sale}</td>
+                                <td>${product.categories.map(category => category.name).join('<br>')}</td>
+                                <td>
+                                    <div style="display: flex;">
+                                        <button type="button" class="btn btn-success" style="margin-left: 5px;" data-bs-toggle="modal" data-bs-target="#showModal" onclick="displayShowModal(${product.id})">
+                                            <i class="fas fa-eye"></i>
+                                        </button>
+                                        ${roleUsers.some(role => ['admin','super-admin'].includes(role)) ? `
+                                            <button type="button" class="btn btn-warning" style="margin-left: 5px;" data-bs-toggle="modal" data-bs-target="#editModal" onclick="displayEditModal(${product.id})">
+                                                <i class="fas fa-edit"></i>
+                                            </button>` : ''
+                                        }
+                                        ${roleUsers.some(role => ['super-admin'].includes(role)) ? `
+                                            <button type="button" class="btn btn-danger" style="margin-left: 5px;" onclick="handleDelete(${product.id})">
+                                                <i class="fas fa-trash-alt"></i>
+                                            </button>` : ''
+                                        }
+                                        
+                                    </div>
+                                </td>
+                            </tr>
+                        `;
+                    
+                        $('#searchProductsIndex').append(productHTML); // show du lieu len view index
+                    }
+                }else {
+                    $('#searchProductsIndex').append('<tr><td colspan="7" style="text-align: center;">Không có kết quả</td></tr>');
+                }
             }
             
-        },
+        },error:function(error) {
+            alert('Error ' + error.message);
+        }
     });
 }
